+++ include/dlaf/factorization/internal.h
@@ -1,0 +1,19 @@
+//
+// Distributed Linear Algebra with Future (DLAF)
+//
+// Copyright (c) 2018-2019, ETH Zurich
+// All rights reserved.
+//
+// Please, refer to the LICENSE file in the root directory.
+// SPDX-License-Identifier: BSD-3-Clause
+//
+#pragma once
+
+#include "dlaf/types.h"
+
+namespace dlaf {
+
+template <Backend execution>
+struct Factorization;
+
+}
