+++ include/dlaf/tile.h
@@ -183,8 +183,8 @@ private:
 
 /// @brief Create a common::Buffer from a Tile
 template <class T, Device device>
-auto create_buffer(const dlaf::Tile<T, device>& tile) {
-  return dlaf::common::Buffer<T*>(tile.ptr({0, 0}), tile.size().cols(), tile.size().rows(), tile.ld());
+auto create_buffer(const Tile<T, device>& tile) {
+  return common::Buffer<T*>(tile.ptr({0, 0}), tile.size().cols(), tile.size().rows(), tile.ld());
 }
 
 #include <dlaf/tile.tpp>
