+++ include/dlaf/types.h
@@ -18,6 +18,8 @@ using SizeType = int;
 
 enum class Device { CPU, GPU };
 
+enum class Backend { MC, GPU };
+
 template <class T>
 struct TypeInfo;
