+++ test/unit/factorization/mc/test_cholesky.cpp
@@ -7,7 +7,7 @@
 // Please, refer to the LICENSE file in the root directory.
 // SPDX-License-Identifier: BSD-3-Clause
 //
-#include "dlaf/mc/cholesky.h"
+#include "dlaf/factorization/mc.h"
 
 #include "gtest/gtest.h"
 #include "dlaf/communication/communicator_grid.h"
@@ -42,9 +42,7 @@ public:
 TYPED_TEST_SUITE(CholeskyDistributedTest, MatrixElementTypes);
 
 std::vector<LocalElementSize> square_sizes({{10, 10}, {25, 25}, {12, 12}, {0, 0}});
-std::vector<LocalElementSize> rectangular_sizes({{10, 20}, {50, 20}, {0, 10}, {20, 0}});
 std::vector<TileElementSize> square_block_sizes({{3, 3}, {5, 5}});
-std::vector<TileElementSize> rectangular_block_sizes({{10, 30}, {20, 10}});
 
 GlobalElementSize globalTestSize(const LocalElementSize& size) {
   return {size.rows(), size.cols()};
@@ -87,42 +85,10 @@ TYPED_TEST(CholeskyLocalTest, Correctness) {
       Matrix<TypeParam, Device::CPU> mat(size, block_size);
       set(mat, el);
 
-      cholesky(blas::Uplo::Lower, mat);
+      Factorization<Backend::MC>::cholesky(blas::Uplo::Lower, mat);
 
       CHECK_MATRIX_NEAR(res, mat, 4 * (mat.size().rows() + 1) * TypeUtilities<TypeParam>::error,
                         4 * (mat.size().rows() + 1) * TypeUtilities<TypeParam>::error);
-    }
-  }
-}
-
-TYPED_TEST(CholeskyLocalTest, MatrixNotSquareException) {
-  for (const auto& size : rectangular_sizes) {
-    for (const auto& block_size : square_block_sizes) {
-      Matrix<TypeParam, Device::CPU> mat(size, block_size);
-
-      EXPECT_THROW(cholesky(blas::Uplo::Lower, mat), std::invalid_argument);
-    }
-  }
-}
-
-TYPED_TEST(CholeskyLocalTest, BlockNotSquareException) {
-  for (const auto& size : square_sizes) {
-    for (const auto& block_size : rectangular_block_sizes) {
-      Matrix<TypeParam, Device::CPU> mat(size, block_size);
-
-      EXPECT_THROW(cholesky(blas::Uplo::Lower, mat), std::invalid_argument);
-    }
-  }
-}
-
-TYPED_TEST(CholeskyLocalTest, MatrixNotLocalException) {
-  for (const auto& size : square_sizes) {
-    for (const auto& block_size : square_block_sizes) {
-      GlobalElementSize sz = globalTestSize(size);
-      Distribution distribution(sz, block_size, {2, 1}, {0, 0}, {0, 0});
-      Matrix<TypeParam, Device::CPU> mat(std::move(distribution));
-
-      EXPECT_THROW(cholesky(blas::Uplo::Lower, mat), std::invalid_argument);
     }
   }
 }
@@ -169,76 +135,10 @@ TYPED_TEST(CholeskyDistributedTest, Correctness) {
         Matrix<TypeParam, Device::CPU> mat(std::move(distribution));
         set(mat, el);
 
-        cholesky(comm_grid, blas::Uplo::Lower, mat);
+        Factorization<Backend::MC>::cholesky(comm_grid, blas::Uplo::Lower, mat);
 
         CHECK_MATRIX_NEAR(res, mat, 4 * (mat.size().rows() + 1) * TypeUtilities<TypeParam>::error,
                           4 * (mat.size().rows() + 1) * TypeUtilities<TypeParam>::error);
-      }
-    }
-  }
-}
-
-TYPED_TEST(CholeskyDistributedTest, MatrixNotSquareException) {
-  for (const auto& comm_grid : this->commGrids()) {
-    for (const auto& size : rectangular_sizes) {
-      for (const auto& block_size : square_block_sizes) {
-        Index2D src_rank_index(std::max(0, comm_grid.size().rows() - 1),
-                               std::min(1, comm_grid.size().cols() - 1));
-        GlobalElementSize sz = globalTestSize(size);
-        Distribution distribution(sz, block_size, comm_grid.size(), comm_grid.rank(), src_rank_index);
-        Matrix<TypeParam, Device::CPU> mat(std::move(distribution));
-
-        EXPECT_THROW(cholesky(comm_grid, blas::Uplo::Lower, mat), std::invalid_argument);
-      }
-    }
-  }
-}
-
-TYPED_TEST(CholeskyDistributedTest, BlockNotSquareException) {
-  for (const auto& comm_grid : this->commGrids()) {
-    for (const auto& size : square_sizes) {
-      for (const auto& block_size : rectangular_block_sizes) {
-        Index2D src_rank_index(std::max(0, comm_grid.size().rows() - 1),
-                               std::min(1, comm_grid.size().cols() - 1));
-        GlobalElementSize sz = globalTestSize(size);
-        Distribution distribution(sz, block_size, comm_grid.size(), comm_grid.rank(), src_rank_index);
-        Matrix<TypeParam, Device::CPU> mat(std::move(distribution));
-
-        EXPECT_THROW(cholesky(comm_grid, blas::Uplo::Lower, mat), std::invalid_argument);
-      }
-    }
-  }
-}
-
-TYPED_TEST(CholeskyDistributedTest, MatrixNotDistributedOnGridException) {
-  // Test: matrix not distributed according to the given communicator grid
-  for (const auto& comm_grid : this->commGrids()) {
-    for (const auto& size : square_sizes) {
-      for (const auto& block_size : square_block_sizes) {
-        GlobalElementSize sz = globalTestSize(size);
-
-        {
-          // Different grid size
-          Size2D grid_distributed(comm_grid.size().rows() + 1, comm_grid.size().cols() + 1);
-          Distribution distribution(sz, block_size, grid_distributed, comm_grid.rank(), {0, 0});
-          Matrix<TypeParam, Device::CPU> mat(std::move(distribution));
-
-          EXPECT_THROW(cholesky(comm_grid, blas::Uplo::Lower, mat), std::invalid_argument);
-        }
-
-        {
-          // Different rank
-          Index2D rank_index_distributed(std::max(0, comm_grid.rank().row() - 1),
-                                         std::max(0, comm_grid.rank().col() - 1));
-
-          Distribution distribution(sz, block_size, comm_grid.size(), rank_index_distributed, {0, 0});
-          Matrix<TypeParam, Device::CPU> rank(std::move(distribution));
-
-          if (rank.distribution().rankIndex() == comm_grid.rank())
-            continue;
-          else
-            EXPECT_THROW(cholesky(comm_grid, blas::Uplo::Lower, rank), std::invalid_argument);
-        }
       }
     }
   }
